test(dht): improve DhtPacket tests
Verify that:
 - decrypting of payload can't be done with wrong SK
 - contents of struct fields when creating DhtPacket
 - malformed DhtPacket with PingResp can't be parsed as PingReq
